Extract seed-generation pipeline out of submissions.ts
Moves round-0 draft generation, its cost ledger, the creator-initiated
regenerate-seed flow, and the status-page seed preview into a new
apps/api/src/creation/seed-pipeline.ts module (createSeedPipeline factory).

dispatchBuild stays in submissions.ts unchanged — it destructures
seedDeliveryFor/seedBuild/regenerateSeed/publishSeedPreview off the
constructed seedPipeline instance, so its own body needed no edits.
builderOf/backendFor/store/now/gameSeeder/seedAvailabilityGate/githubClient/
publishedRef are passed in as options, same injected-option pattern as the
prior 9 batches.

submissions.ts drops from 4217 to 4032 lines; the new file is 233 lines,
0 comment-prose debt.
